refactor: move atom functions to the right place
