+++ lib/helpers.js
@@ -1,5 +1,4 @@
 import log from './logger';
-import getAtom from './atoms';
 import _ from 'lodash';
 import B from 'bluebird';
 import { errorFromMJSONWPStatusCode } from 'appium-base-driver';
@@ -172,37 +171,6 @@ function checkParams (params) {
   }
 }
 
-async function wrapScriptForFrame (script, frame) {
-  log.debug(`Wrapping script for frame '${frame}'`);
-  let elFromCache = await getAtom('get_element_from_cache');
-  return `(function (window) { var document = window.document; ` +
-         `return (${script}); })((${elFromCache.toString('utf8')})(${JSON.stringify(frame)}))`;
-}
-
-async function getScriptForAtom (atom, args, frames, asyncCallBack = null) {
-  let atomSrc = await getAtom(atom);
-  let script;
-  if (frames.length > 0) {
-    script = atomSrc;
-    for (const frame of frames) {
-      script = await wrapScriptForFrame(script, frame);
-    }
-  } else {
-    log.debug(`Executing '${atom}' atom in default context`);
-    script = `(${atomSrc})`;
-  }
-
-  // add the arguments, as strings
-  args = args.map(JSON.stringify);
-  if (asyncCallBack) {
-    script += `(${args.join(',')}, ${asyncCallBack}, true )`;
-  } else {
-    script += `(${args.join(',')})`;
-  }
-
-  return script;
-}
-
 function simpleStringify (value, multiline = false) {
   if (!value) {
     return JSON.stringify(value);
@@ -266,6 +234,6 @@ function convertResult (res) {
 
 export {
   appInfoFromDict, pageArrayFromDict, getDebuggerAppKey,
-  getPossibleDebuggerAppKeys, checkParams, wrapScriptForFrame, getScriptForAtom,
-  simpleStringify, deferredPromise, convertResult, RESPONSE_LOG_LENGTH,
+  getPossibleDebuggerAppKeys, checkParams, simpleStringify, deferredPromise,
+  convertResult, RESPONSE_LOG_LENGTH,
 };
