+++ lib/remote-debugger.js
@@ -4,9 +4,10 @@ import { errors } from 'appium-base-driver';
 import { RpcClientSimulator } from './rpc';
 import messageHandlers from './message-handlers';
 import { appInfoFromDict, pageArrayFromDict, getDebuggerAppKey,
-         getPossibleDebuggerAppKeys, checkParams, getScriptForAtom,
+         getPossibleDebuggerAppKeys, checkParams,
          simpleStringify, deferredPromise, convertResult,
          RESPONSE_LOG_LENGTH } from './helpers';
+import { getScriptForAtom } from './atoms';
 import { util, timing } from 'appium-support';
 import { retryInterval, waitForCondition } from 'asyncbox';
 import _ from 'lodash';
